pyproject.toml
[tool.poetry]
name = "basic-foundation"
version = "0.1.0"
description = "Opinionated python based Saas platform"
authors = ["phernandez <phernandez@basicmachines.co>"]
license = "Apache-2.0"
readme = "README.md"

packages = [
    { include = "basic_foundation" }, # core components
    { include = "modules.basic_foundation.api", from = "modules" }, # API module
    { include = "modules.basic_foundation.web", from = "modules" }   # Web module
]

[tool.poetry.dependencies]
python = "3.12.1"
ruff = "^0.2.1"
setuptools = "^69.0.3"
loguru = "^0.7.2"
fastapi-jwt = { extras = ["authlib"], version = "^0.3.0" }
typer = "^0.12.3"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
psycopg2 = "^2.9.9"
emails = "^0.6"
aiosmtpd = "^1.4.6"
mockito = "^1.5.0"
fastapi = { extras = ["standard"], version = "^0.112.0" }
pyjwt = "^2.9.0"
itsdangerous = "^2.2.0"
starlette-wtf = "^0.4.5"
pytest-cov = "^5.0.0"
pyright = "^1.1.381"

[tool.poetry.group.app.dependencies]
uvicorn = "^0.25.0"
jinja2 = "^3.1.2"
aiosqlite = "^0.19.0"
pydantic-settings = "^2.1.0"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.25" }
asyncpg = "^0.29.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"
coverage = "^7.4.0"
playwright = "^1.41.2"
pytest-playwright = "^0.4.4"

[tool.poetry.group.ci.dependencies]
pytest-asyncio = "^0.23.3"
asgi-lifespan = "^2.1.0"
httpx-oauth = "^0.13.1"
pytest-mock = "^3.12.0"

[tool.poetry.group.github-actions]
optional = true

# only used in github actions
[tool.poetry.group.github-actions.dependencies]
pytest-github-actions-annotate-failures = "^0.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.app"


[tool.pytest.ini_options]
pythonpath = [
    ".", "./app/src",
    ".", "./app/tests",
]
filterwarnings = [
    "ignore::DeprecationWarning:",
]

markers = [
    "playwright: playwright web tests, requires a running app instance",
]

[tool.pyright]
include = ["app"]
exclude = ["**/node_modules",
    "**/__pycache__",
]
ignore = ["./app/tests"]
defineConstant = { DEBUG = true }
stubPath = "app/src/stubs"

reportMissingImports = "error"
reportMissingTypeStubs = false

pythonVersion = "3.12"
